Refactor transform operator shared helpers
